
\name{poisson.gc1}
\alias{poisson.gc1}
\title{The Poisson Marginal of Class \code{\link{marginal.gc1}}}
\usage{
poisson.gc1(lambda = 1)
}
\arguments{
\item{lambda}{a non-negative scalar of the mean parameter.}
}
\value{
An object of class \code{\link{marginal.gc1}} representing the marginal component.
}
\description{
The Poisson marginal used for simulation and computing correlation in
trans-Gaussian random field.
}
\author{
Zifei Han \email{hanzifei1@gmail.com}
}
\seealso{
\code{\link{marginal.gc1}},  \code{\link{beta.gc1}},
          \code{\link{binomial.gc1}}, \code{\link{gm.gc1}},
          \code{\link{gaussian.gc1}}, \code{\link{negbin.gc1}},
          \code{\link{weibull.gc1}}, \code{\link{zip.gc1}}
}

\keyword{Marginal}
\keyword{Simulation}
