\name{summary.mlegc}
\alias{print.mlegc}
\alias{summary.mlegc}
\alias{print.summary.mlegc}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for Extracting Information from Fitted Object of Class \code{mlegc}}


\usage{
\method{summary}{mlegc}(object, \dots)
}

\description{
	Return a summary table of results from model fitting.
}
%- maybe also 'usage' for other objects documented here.
\arguments{
 \item{object}{an object of class \code{mlegc} inherited from function \code{\link{mlegc}}.
}

\item{\dots}{additional arguments, but currently not used.}

}

\value{
A table summary of the estimates, standard error, z-value and several information criteria.
}


\author{
Zifei Han \email{hanzifei1@gmail.com}
}


\seealso{
\code{\link{mlegc}}
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{Estimation}% use one of  RShowDoc("KEYWORDS")
